+++ order/src/main/kotlin/com/example/orderservice/vo/OrderItem.kt
@@ -6,7 +6,7 @@ class OrderItem(
 	var productId: String? = null,
 	var qty: Int? = null,
 	var unitPrice: Int? = null,
-	var totalPrice: Int? = null,
+	var totalPrice: Long? = null,
 	var orderId: String? = null,
 	var userId: String? = null,
 	var createdAt: LocalDateTime? = null,
